I think I changed things

+++ ThunderSurge-core/daybreak/components/Skybox.h
@@ -14,7 +14,7 @@ namespace daybreak {
 			GLuint m_id, m_vao, m_vbo;
 			Shader* m_shader;
 
-			void load(vector<string> faces, int size) {
+			void load(vector<string> faces, float size) {
 
 				float cubemap[] = {     
 					-size,  size, -size,
@@ -93,7 +93,7 @@ namespace daybreak {
 
 			}
 		public:
-			Skybox(int size, string right, string left, string top, string bottom, string front, string back) {
+			Skybox(float size, string right, string left, string top, string bottom, string front, string back) {
 				load(vector<string> { right, left, top, bottom, front, back }, size);
 				m_shader = new Shader("daybreak/graphics/shaders/skybox.vert", "daybreak/graphics/shaders/skybox.frag", NULL, false);
 				m_shader->enable();
